i18n: Remove XML entity for docs URL. (#1052)
The &folder_ignore_delete_docs_url; entity is used to avoid repeating
the URL verbatim in translations strings.  While this is a good
approach in principle, it does come with a couple of drawbacks:

1. Automated processing via XSLT will resolve the entity refs
automatically, this cannot be easily avoided.

2. It is a one-off pattern, used only for this URL.  All others are
part of the translation strings verbatim.  If used stand-alone, they
are marked translatable="false".

3. It is still error-prone to typos when making the translation.  If
the URL is screwed up, it won't get the user to its target.  But if
the entity reference is misspelled, there will be nothing to look at
probably.

Any possible benefits don't outweigh these and the added complexity,
thus get rid of the entity references (using a simple XSLT transform).