chore/fix: handle per-folder databases properly
This kills two birds with one stone, by fixing the problem where it
doesn't back up the WAL file, and making so it is compatible with
versions of Syncthing with syncthing/syncthing#10042 applied
(v2.0.0-beta.6 and newer).
Note that this commit is *only* compatible with those versions.

In this case I bumped it to v2.0.0-beta.10, but any version
v2.0.0-beta.6 or newer will do.

This backs up everything in the index-v2 folder (including the WAL file
[if it exists])

According to https://sqlite.org/wal.html:

```quote
The WAL file is part of the persistent state of the database and should be kept with the database if the database is copied or moved.
```