Update strings.xml
This traslation should be fixed (I speak native italian). The current translastion is misleading, because the original word "revert" in this context has the meaning of "overwrite": the local content of the folder will be replaced by the remote one. 

So this cannot be translated with "ripristina" wich literally means "restore". This makes the button misleading, the opposite of the actual meaning. Basically the current translation is literally "restore local changes", which is not what the buttons means.

I propose "sovrascrivi", which means "overwrite". 
So in the end: "overwrite local changes".